requirements.txt
cs50
Flask
Flask-Session
requests
gunicorn
plotly.graph_objs
numpy
datatime
dateutil